+++ model/package.go
@@ -17,7 +17,7 @@ var FreePackage = Package{
 
 // Package 套餐模型
 type Package struct {
-	Id            uint      `json:"id" gorm:"primaryKey"`
+	Id            int64     `json:"id" gorm:"primaryKey"`
 	Name          string    `json:"name" gorm:"type:varchar(50);not null"`
 	Description   string    `json:"description" gorm:"type:varchar(255)"`
 	Price         float64   `json:"price" gorm:"type:decimal(10,2);not null"`
@@ -53,9 +53,9 @@ func (p *Package) ToResponsePackage() map[string]interface{} {
 
 // Subscription 订阅模型
 type Subscription struct {
-	Id          uint   `json:"id" gorm:"primaryKey"`
-	UserId      uint   `json:"user_id" gorm:"not null;index"`
-	PackageId   uint   `json:"package_id" gorm:"not null;index"`
+	Id          int64  `json:"id" gorm:"primaryKey"`
+	UserId      int64  `json:"user_id" gorm:"not null;index"`
+	PackageId   int64  `json:"package_id" gorm:"not null;index"`
 	Status      string `json:"status" gorm:"type:varchar(20);not null;default:'active'"` // active, expired, cancelled
 	StartDate   int64  `json:"start_date" gorm:"not null"`
 	ExpiryDate  int64  `json:"expiry_date" gorm:"not null"`
@@ -67,10 +67,10 @@ type Subscription struct {
 
 // TokenDistribution 每月Token分发记录
 type TokenDistribution struct {
-	Id             uint  `json:"id" gorm:"primaryKey"`
-	UserId         uint  `json:"user_id" gorm:"not null;index"`
-	SubscriptionId uint  `json:"subscription_id" gorm:"index"`
-	PackageId      uint  `json:"package_id" gorm:"index"`
+	Id             int64 `json:"id" gorm:"primaryKey"`
+	UserId         int64 `json:"user_id" gorm:"not null;index"`
+	SubscriptionId int64 `json:"subscription_id" gorm:"index"`
+	PackageId      int64 `json:"package_id" gorm:"index"`
 	Amount         int   `json:"amount" gorm:"not null"`
 	DistributedAt  int64 `json:"distributed_at" gorm:"not null"`
 	CreatedAt      int64 `json:"created_at"`
